refactor(watcher): split watcher classes
